shadow-dom/events/event-dispatch/test-003.html

<!DOCTYPE html>
<!--
Distributed under both the W3C Test Suite License [1] and the W3C
3-clause BSD License [2]. To contribute to a W3C Test Suite, see the
policies and contribution forms [3].

[1] http://www.w3.org/Consortium/Legal/2008/04-testsuite-license
[2] http://www.w3.org/Consortium/Legal/2008/03-bsd-license
[3] http://www.w3.org/2004/10/27-testcases
-->
<html>
<head>
<title>Shadow DOM Test: A_05_05_03</title>
<link rel="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<link rel="help" href="https://w3c.github.io/webcomponents/spec/shadow/#event-path-trimming">
<meta name="assert" content="Event Path Trimming: event listeners for trusted events must not be invoked on a node for which the target and relatedTarget are the same.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../../testcommon.js"></script>
<link rel="stylesheet" href="/resources/testharness.css">
</head>
<body>
<div id="log"></div>
<script>
var A_05_05_03_T01 = async_test('A_05_05_03_T01');

A_05_05_03_T01.step(unit(function (ctx) {

    var d = newRenderedHTMLDocument(ctx);

    var host = d.createElement('div');
    host.setAttribute('style', 'height:50%; width:100%');
    host.setAttribute('id', 'host');
    d.body.appendChild(host);

    //Shadow root to play with
    var s = host.createShadowRoot();

    var div1 = d.createElement('div');
    div1.setAttribute('style', 'height:50%; width:100%');
    div1.setAttribute('id', 'div1');
    s.appendChild(div1);

    var div2 = d.createElement('div');
    div2.setAttribute('style', 'height:50%; width:100%');
    div2.setAttribute('id', 'div2');
    s.appendChild(div2);

    // Add an event listner above its shadow root. If both target and
    // relatedTarget are in the same shadow tree under the root,
    // retargeted target and relatedTarget will be the same node
    // and the event must not invoke the handler.
    host.addEventListener('mouseover', A_05_05_03_T01.step_func(function(event) {
        if (('isTrusted' in event) && event.isTrusted) {
            assert_true(false, 'Trusted event must not invoke its listeners' +
                'if target and relatedTarget are the same');
        }
    }), false);

    // Note: this is not a "trusted event", thus user agents may not
    // suppress invoking event handlers.
    var evt = document.createEvent("MouseEvents");
    evt.initMouseEvent("mouseover", true, false, window,
      0, 10, 10, 10, 10, false, false, false, false, 0, div2);

    div1.dispatchEvent(evt);

    A_05_05_03_T01.done();
}));
</script>
</body>
</html>
